nuttx write call cannot handle 64bit signed integer, circumvent it via params[0]

+++ arch/x86_64/include/intel64/irq.h
@@ -193,7 +193,7 @@ struct xcptcontext
   uint32_t is_linux;
   int32_t linux_sock;
   uint64_t linux_tcb;
-  int syscall_ret;
+  uint64_t syscall_ret;
 
   void* __brk;
   void* __min_brk;
